tambah siswa

+++ database/migrations/2026_05_12_053948_add_photo_to_users_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->string('photo')->nullable()->after('password');
-        });
+        if (!Schema::hasColumn('users', 'photo')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->string('photo')->nullable()->after('password');
+            });
+        }
     }
 
     /**
@@ -21,8 +23,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('photo');
-        });
+        if (Schema::hasColumn('users', 'photo')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('photo');
+            });
+        }
     }
 };
